+++ package.json
@@ -2,22 +2,24 @@
   "name": "home_work_3",
   "version": "1.0.0",
   "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "start": "webpack-dev-server --mode development",
+    "build": "webpack --mode production",
+    "lint": "tslint src/**/*.ts --fix"
+  },
   "dependencies": {
+    "hls.js": "^0.11.0",
     "normalize.css": "^8.0.0",
     "pepjs": "^0.4.3"
   },
   "devDependencies": {
-    "@babel/core": "^7.1.2",
-    "@babel/preset-env": "^7.1.0",
-    "babel-loader": "^8.0.4",
-    "babel-polyfill": "^6.26.0",
+    "@types/hls.js": "^0.10.2",
+    "@types/webaudioapi": "0.0.27",
     "css-loader": "^1.0.0",
-    "eslint": "^5.6.1",
-    "eslint-config-airbnb-base": "^13.1.0",
-    "eslint-loader": "^2.1.1",
-    "eslint-plugin-import": "^2.14.0",
     "file-loader": "^2.0.0",
     "mini-css-extract-plugin": "^0.4.3",
+    "promise-polyfill": "^8.1.0",
     "style-loader": "^0.23.0",
     "ts-loader": "^5.2.2",
     "tslint": "^5.11.0",
@@ -26,12 +28,6 @@
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.9",
     "whatwg-fetch": "^3.0.0"
-  },
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "webpack-dev-server --mode development",
-    "build": "webpack --mode production",
-    "lint": "tslint src/**/*.ts --fix"
   },
   "keywords": [],
   "author": "",
