+++ public/video.html
@@ -8,7 +8,6 @@
   <meta name="robots" content="noindex">
   <title>Умный дом — Видеонаблюдение</title>
   <link rel="stylesheet" href="style.css">
-  <script src="https://cdn.jsdelivr.net/npm/hls.js@latest"></script>
 </head>
 
 <body>
